// frontend/src/client/types.gen.ts
export type FeatureResponse = {
  /**
   * Name
   */
  name: string;
  /**
   * Code
   */
  code: string;
  /**
   * Unit Price
   */
  unit_price: string;
  /**
   * Max Unit Limit
   */
  max_unit_limit: number;
  /**
   * Is Active
   */
  is_active?: boolean;
  /**
   * Id
   */
  id: number;
};
// ...
export type PlanResponse = {
  /**
   * Name
   */
  name: string;
  /**
   * Monthly Fee
   *
   * The recurring monthly price charged for this plan
   */
  monthly_fee: number;
  /**
   * Id
   */
  id: number;
  /**
   * Features
   */
  features?: Array<FeatureResponse>;
};

/**
 * SubscriptionCreate
 */
export type SubscriptionCreate = {
  /**
   * Plan Id
   *
   * ID of the plan to subscribe to
   */
  plan_id: number;
};

/**
 * SubscriptionResponse
 */
export type SubscriptionResponse = {
  /**
   * Id
   */
  id: number;
  /**
   * User Id
   */
  user_id: number;
  /**
   * Plan Id
   */
  plan_id: number;
  /**
   * Billing Day
   */
  billing_day: number;
  /**
   * Last Billing Date
   */
  last_billing_date?: string | null;
  /**
   * Is Active
   */
  is_active: boolean;
  plan: PlanResponse;
};
// ...
export type UsageResponse = {
  /**
   * Id
   */
  id: number;
  /**
   * Subscription Id
   */
  subscription_id: number;
  /**
   * Feature Id
   */
  feature_id: number;
  /**
   * Units Used
   */
  units_used: number;
  feature?: FeatureResponse | null;
};
// ...
export type UserResponse = {
  /**
   * Name
   */
  name: string;
  /**
   * Email
   */
  email: string;
  role?: UserRole;
  /**
   * Id
   */
  id: number;
  /**
   * Profile Img
   */
  profile_img?: string | null;
};

/**
 * UserRole
 */
export type UserRole = "admin" | "buyer";